+++ aesara/link/jax/dispatch/elemwise.py
@@ -3,7 +3,7 @@ import jax.numpy as jnp
 
 from aesara.link.jax.dispatch.basic import jax_funcify, jnp_safe_copy
 from aesara.tensor.elemwise import CAReduce, DimShuffle, Elemwise
-from aesara.tensor.math import LogSoftmax, Softmax, SoftmaxGrad
+from aesara.tensor.special import LogSoftmax, Softmax, SoftmaxGrad
 
 
 @jax_funcify.register(Elemwise)
