+++ aesara/tensor/rewriting/__init__.py
@@ -3,5 +3,6 @@ import aesara.tensor.rewriting.elemwise
 import aesara.tensor.rewriting.extra_ops
 import aesara.tensor.rewriting.math
 import aesara.tensor.rewriting.shape
+import aesara.tensor.rewriting.special
 import aesara.tensor.rewriting.subtensor
 import aesara.tensor.rewriting.uncanonicalize
