+++ tests/tensor/test_math.py
@@ -8,9 +8,7 @@ from itertools import product
 import numpy as np
 import pytest
 from numpy.testing import assert_array_equal
-from scipy.special import log_softmax as scipy_log_softmax
 from scipy.special import logsumexp as scipy_logsumexp
-from scipy.special import softmax as scipy_softmax
 
 import aesara.scalar as aes
 from aesara.compile.debugmode import DebugMode
@@ -36,14 +34,11 @@ from aesara.tensor.elemwise import CAReduce, Elemwise
 from aesara.tensor.math import (
     Argmax,
     Dot,
-    LogSoftmax,
     MatMul,
     MaxAndArgmax,
     Mean,
     Prod,
     ProdWithoutZeros,
-    Softmax,
-    SoftmaxGrad,
     Sum,
     _allclose,
     _dot,
@@ -84,7 +79,6 @@ from aesara.tensor.math import (
     log1p,
     log2,
     log10,
-    log_softmax,
     logaddexp,
     logsumexp,
     matmul,
@@ -110,7 +104,6 @@ from aesara.tensor.math import (
     sin,
     sinh,
     smallest,
-    softmax,
     sqr,
     sqrt,
     sub,
@@ -3528,129 +3521,3 @@ class TestMatMul(utt.InferShapeTester):
                 [x1, x2],
                 self.op_class,
             )
-
-
-class TestSoftmax(utt.InferShapeTester):
-    @pytest.mark.parametrize("axis", [None, 0, 1, 2, 3, -1, -2])
-    def test_perform(self, axis):
-        x = tensor4("x")
-        rng = np.random.default_rng(utt.fetch_seed())
-        xv = rng.standard_normal((2, 3, 4, 5)).astype(config.floatX)
-
-        f = function([x], softmax(x, axis=axis))
-        assert np.allclose(f(xv), scipy_softmax(xv, axis=axis))
-
-    @pytest.mark.parametrize("column", [0, 1, 2, 3])
-    @pytest.mark.parametrize("axis", [None, 0, 1])
-    def test_grad(self, axis, column):
-        def f(a):
-            return softmax(a, axis=axis)[:, column]
-
-        rng = np.random.default_rng(utt.fetch_seed())
-        utt.verify_grad(f, [rng.random((3, 4, 2))])
-
-    def test_infer_shape(self):
-        admat = matrix()
-        rng = np.random.default_rng(utt.fetch_seed())
-        admat_val = rng.random((3, 4)).astype(config.floatX)
-        self._compile_and_check(
-            [admat], [Softmax(axis=-1)(admat)], [admat_val], Softmax
-        )
-
-    def test_vector_perform(self):
-        x = vector()
-        f = function([x], softmax(x, axis=None))
-
-        rng = np.random.default_rng(utt.fetch_seed())
-        xv = rng.standard_normal((6,)).astype(config.floatX)
-        assert np.allclose(f(xv), scipy_softmax(xv))
-
-    def test_vector_grad(self):
-        def f(a):
-            return softmax(a, axis=None)
-
-        rng = np.random.default_rng(utt.fetch_seed())
-        utt.verify_grad(f, [rng.random((4))])
-
-    def test_valid_axis(self):
-        with pytest.raises(TypeError):
-            Softmax(1.5)
-
-        x = [tensor3()] * LogSoftmax.nin
-        Softmax(2)(*x)
-        Softmax(-3)(*x)
-
-        with pytest.raises(ValueError):
-            Softmax(3)(*x)
-
-        with pytest.raises(ValueError):
-            Softmax(-4)(*x)
-
-
-class TestLogSoftmax(utt.InferShapeTester):
-    @pytest.mark.parametrize("column", [0, 1, 2, 3])
-    @pytest.mark.parametrize("axis", [None, 0, 1])
-    def test_matrix_grad(self, axis, column):
-        def f(a):
-            return log_softmax(a, axis=axis)[:, column]
-
-        rng = np.random.default_rng(utt.fetch_seed())
-        utt.verify_grad(f, [rng.random((3, 4))])
-
-    def test_vector_perform(self):
-        x = vector()
-        f = function([x], log_softmax(x, axis=None))
-
-        rng = np.random.default_rng(utt.fetch_seed())
-        xv = rng.standard_normal((6,)).astype(config.floatX)
-        assert np.allclose(f(xv), scipy_log_softmax(xv))
-
-    def test_vector_grad(self):
-        def f(a):
-            return log_softmax(a, axis=None)
-
-        rng = np.random.default_rng(utt.fetch_seed())
-        utt.verify_grad(f, [rng.random((4,))])
-
-    def test_valid_axis(self):
-        with pytest.raises(TypeError):
-            LogSoftmax(1.5)
-
-        x = [tensor3()] * LogSoftmax.nin
-        LogSoftmax(2)(*x)
-        LogSoftmax(-3)(*x)
-
-        with pytest.raises(ValueError):
-            LogSoftmax(3)(*x)
-
-        with pytest.raises(ValueError):
-            LogSoftmax(-4)(*x)
-
-
-class TestSoftmaxGrad(utt.InferShapeTester):
-    def test_infer_shape(self):
-        admat = matrix()
-        bdmat = matrix()
-        rng = np.random.default_rng(utt.fetch_seed())
-        admat_val = rng.random((3, 4)).astype(config.floatX)
-        bdmat_val = rng.random((3, 4)).astype(config.floatX)
-        self._compile_and_check(
-            [admat, bdmat],
-            [SoftmaxGrad(axis=-1)(admat, bdmat)],
-            [admat_val, bdmat_val],
-            SoftmaxGrad,
-        )
-
-    def test_valid_axis(self):
-        with pytest.raises(TypeError):
-            SoftmaxGrad(1.5)
-
-        x = [tensor3()] * SoftmaxGrad.nin
-        SoftmaxGrad(2)(*x)
-        SoftmaxGrad(-3)(*x)
-
-        with pytest.raises(ValueError):
-            SoftmaxGrad(3)(*x)
-
-        with pytest.raises(ValueError):
-            SoftmaxGrad(-4)(*x)
